+++ lib/entities/request/index.ts
@@ -1,4 +1,3 @@
-import { Protocol } from '@uniswap/router-sdk';
 import { TradeType } from '@uniswap/sdk-core';
 import { BigNumber } from 'ethers';
 
@@ -144,17 +143,4 @@ export function defaultRequestKey(request: QuoteRequest): string {
     amount: info.amount.toString(),
     type: info.type,
   });
-}
-
-export function parseProtocol(protocol: string): Protocol {
-  switch (protocol.toLowerCase()) {
-    case 'v2':
-      return Protocol.V2;
-    case 'v3':
-      return Protocol.V3;
-    case 'mixed':
-      return Protocol.MIXED;
-    default:
-      throw new Error(`Invalid protocol: ${protocol}`);
-  }
 }
